pyproject.toml
[build-system]
requires = [
  "hatchling",
  "hatch-zipped-directory",
]
build-backend = "hatchling.build"

[project]
name = "elmo-iess-alarm"
dynamic = ["version"]
description = "Home Assistant integration that provides a full-fledged Alarm Panel to control your Elmo/IESS alarm systems."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = [
  "python",
  "home-automation",
  "home-assistant",
  "alarm-system",
  "econnect",
  "elmo",
  "metronet",
  "iess",
]
authors = [
  { name = "Emanuele Palazzetti", email = "emanuele.palazzetti@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "homeassistant",
  "econnect-python==0.6.0",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "pre-commit",
  # Test
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "responses",
  "tox",
  # Home Assistant fixtures
  "freezegun",
  "pytest-asyncio",
  "pytest-socket",
  "requests-mock",
  "syrupy",
]

lint = [
  "black",
  "flake8",
]

all = [
  "elmo-iess-alarm[dev]",
  "elmo-iess-alarm[lint]",
]

[project.urls]
Documentation = "https://github.com/palazzem/ha-econnect-alarm#readme"
Issues = "https://github.com/palazzem/ha-econnect-alarm/issues"
Source = "https://github.com/palazzem/ha-econnect-alarm"

[tool.hatch.version]
path = "custom_components/elmo_iess_alarm/manifest.json"
pattern = '"version":\s"(?P<version>[^"]+)"'

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.sdist]
only-include = ["custom_components/elmo_iess_alarm"]

[tool.hatch.build.targets.zipped-directory]
only-include = ["custom_components/elmo_iess_alarm"]
